TrustEval-toolkit/pyproject.toml
# [build-system]
# requires = ["setuptools>=61.0"]
# build-backend = "setuptools.build_meta"

[project]
name = "trusteval"
version = "0.1.0"
description = "TrustEval: A Dynamic Evaluation Toolkit on the Trustworthiness ofGenerative Foundation Models"
authors = [
    {name="Your Name", email="your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]

# Core dependencies shared by all installations
dependencies = [
    "pandas>=2.2.0",
    "protobuf>=5.0.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.66.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.37.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.11.0",
    "google-generativeai",
    "keybert>=0.8.0",
    "openai>=1.52.0",
    "python-multipart>=0.0.9",
    "streamlit>=1.38.0",
    "zhipuai>=2.1.5",
    "replicate>=0.27.0",
    "accelerate==0.30.1",
    "markdown>=3.7",
    "gdown>=5.2.0",
    "opencv-python>=4.10.0",
]

[project.optional-dependencies]
# Optional dependencies for extended functionality
local = [
    "datasets>=2.15.0",
    "diffusers==0.31.0",
    "Pillow==10.0.0",
    "pydantic==2.9.2",
    "sentencepiece==0.2.0",
    "torch>=2.1.0",
    "transformers>=4.41.2",
]

[project.scripts]
configuration = "trusteval.src.configuration:configuration"

[tool.setuptools]
include-package-data = true
packages = ["trusteval"]

[tool.setuptools.package-data]
trusteval = ['**/*.json', '**/*.yaml', '**/*.py']